quote marks fixes in custom_title
